Move diff-range computation tests
